Explicitly cast an argument to ssprintf representing the number of Debian archives containing a package to 'int', so gcc will be happy about its type. (Closes: #488132)

+++ src/cmdline/cmdline_resolver.cc
@@ -236,7 +236,7 @@ static pkgCache::VerIterator choose_version(const vector<pkgCache::VerIterator> 
 
   for(vector<pkgCache::VerIterator>::size_type i = 0;
       i < choices.size(); ++i)
-    cout << ssprintf(" (%d) %s", i+1, archives_text(choices[i]).c_str()) << endl;
+    cout << ssprintf(" (%d) %s", (int)(i+1), archives_text(choices[i]).c_str()) << endl;
 
   while(1)
     {
